+++ tests/test_initial_configuration_from_server.rs
@@ -1,66 +1,13 @@
 use appconfiguration::{
-    AppConfigurationClient, AppConfigurationClientHttp, ConfigurationId, LiveConfigurationImpl,
-    ServiceAddress, TokenProvider,
+    AppConfigurationClient, AppConfigurationClientHttp, ConfigurationId, ServiceAddress,
 };
 
-use std::io::{BufRead, BufReader, Write};
-use std::net::{TcpListener, TcpStream};
-use std::path::PathBuf;
+use std::net::TcpListener;
 use std::sync::mpsc::channel;
 use std::thread::{sleep, spawn};
 use std::time::Duration;
-use tungstenite::WebSocket;
 
-fn handle_config_request_trivial_config(server: &TcpListener) {
-    let json_payload = serde_json::json!({
-        "environments": [
-            {
-                "name": "Dev",
-                "environment_id": "dev",
-                "features": [],
-                "properties": []
-            }
-        ],
-        "segments": []
-    });
-    handle_config_request(server, json_payload.to_string());
-}
-
-fn handle_config_request_enterprise_example(server: &TcpListener) {
-    let mut mocked_data = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    mocked_data.push("data/data-dump-enterprise-plan-sdk-testing.json");
-    let json_payload = std::fs::read_to_string(mocked_data).unwrap();
-
-    handle_config_request(server, json_payload);
-}
-
-fn handle_config_request(server: &TcpListener, json_payload: String) {
-    let (mut stream, _) = server.accept().unwrap();
-
-    let buf_reader = BufReader::new(&stream);
-    let http_request: Vec<_> = buf_reader
-        .lines()
-        .map(|result| result.unwrap())
-        .take_while(|line| !line.is_empty())
-        .collect();
-    assert_eq!(http_request[0], "GET /test/feature/v1/instances/guid/config?action=sdkConfig&environment_id=dev&collection_id=collection_id HTTP/1.1");
-
-    let response = format!(
-        "HTTP/1.1 200 OK\r\nContent-Length: {}\r\n\r\n{}",
-        json_payload.len(),
-        json_payload
-    );
-    stream.write_all(response.as_bytes()).unwrap();
-}
-
-fn handle_websocket(server: &TcpListener) -> WebSocket<TcpStream> {
-    let (stream, _) = server.accept().unwrap();
-    let mut websocket = tungstenite::accept(stream).unwrap();
-    websocket
-        .send(tungstenite::Message::text("test message".to_string()))
-        .unwrap();
-    websocket
-}
+mod common;
 
 struct ServerHandle {
     _terminator: std::sync::mpsc::Sender<()>,
@@ -75,9 +22,9 @@ fn server_thread() -> ServerHandle {
     let port = server.local_addr().unwrap().port();
     spawn(move || {
         // notify client that config changed
-        let mut websocket = handle_websocket(&server);
+        let mut websocket = common::handle_websocket(&server);
 
-        handle_config_request_enterprise_example(&server);
+        common::handle_config_request_enterprise_example(&server);
 
         // Wait until the client has already tested the first configuration
         update_config_rx.recv().unwrap();
@@ -90,7 +37,7 @@ fn server_thread() -> ServerHandle {
             .unwrap();
 
         // client will request changed config asynchronously
-        handle_config_request_trivial_config(&server);
+        common::handle_config_request_trivial_config(&server);
 
         let _ = receiver.recv();
     });
@@ -98,24 +45,6 @@ fn server_thread() -> ServerHandle {
         _terminator: terminator,
         config_updated: update_config_tx,
         port,
-    }
-}
-
-#[derive(Debug)]
-struct MockTokenProvider {}
-
-impl TokenProvider for MockTokenProvider {
-    fn get_access_token(&self) -> appconfiguration::NetworkResult<String> {
-        Ok("mock_token".into())
-    }
-}
-
-fn wait_until_online(client: &AppConfigurationClientHttp<LiveConfigurationImpl>) {
-    loop {
-        if client.is_online().unwrap() {
-            break;
-        };
-        sleep(Duration::from_millis(10));
     }
 }
 
@@ -134,10 +63,10 @@ fn main() {
         "collection_id".to_string(),
     );
     let client =
-        AppConfigurationClientHttp::new(address, Box::new(MockTokenProvider {}), config_id)
+        AppConfigurationClientHttp::new(address, Box::new(common::MockTokenProvider {}), config_id)
             .unwrap();
 
-    wait_until_online(&client);
+    common::wait_until_online(&client);
 
     let mut features = client.get_feature_ids().unwrap();
     features.sort();
