+++ tests/test_ws_connection_close.rs
@@ -1,52 +1,14 @@
 use appconfiguration::{
-    AppConfigurationClient, AppConfigurationClientHttp, ConfigurationId, LiveConfigurationImpl,
-    ServiceAddress, TokenProvider,
+    AppConfigurationClient, AppConfigurationClientHttp, ConfigurationId, ServiceAddress,
 };
-use tungstenite::WebSocket;
 
-use std::io::{BufRead, BufReader, Write};
-use std::net::{TcpListener, TcpStream};
-use std::path::PathBuf;
+use std::net::TcpListener;
+
 use std::sync::mpsc::channel;
 use std::thread::{sleep, spawn};
 use std::time::Duration;
 
-fn handle_config_request_enterprise_example(server: &TcpListener) {
-    let mut mocked_data = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    mocked_data.push("data/data-dump-enterprise-plan-sdk-testing.json");
-    let json_payload = std::fs::read_to_string(mocked_data).unwrap();
-
-    handle_config_request(server, json_payload);
-}
-
-fn handle_config_request(server: &TcpListener, json_payload: String) {
-    let (mut stream, _) = server.accept().unwrap();
-
-    let buf_reader = BufReader::new(&stream);
-    let http_request: Vec<_> = buf_reader
-        .lines()
-        .map(|result| result.unwrap())
-        .take_while(|line| !line.is_empty())
-        .collect();
-    assert_eq!(http_request[0], "GET /test/feature/v1/instances/guid/config?action=sdkConfig&environment_id=dev&collection_id=collection_id HTTP/1.1");
-
-    let response = format!(
-        "HTTP/1.1 200 OK\r\nContent-Length: {}\r\n\r\n{}",
-        json_payload.len(),
-        json_payload
-    );
-    stream.write_all(response.as_bytes()).unwrap();
-}
-
-fn handle_websocket(server: &TcpListener) -> WebSocket<TcpStream> {
-    let (stream, _) = server.accept().unwrap();
-    let mut websocket = tungstenite::accept(stream).unwrap();
-    websocket
-        .send(tungstenite::Message::text("test message".to_string()))
-        .unwrap();
-
-    websocket
-}
+mod common;
 
 struct ServerHandle {
     _terminator: std::sync::mpsc::Sender<()>,
@@ -61,9 +23,9 @@ fn server_thread() -> ServerHandle {
     let port = server.local_addr().unwrap().port();
     spawn(move || {
         // notify client that config changed
-        let mut websocket = handle_websocket(&server);
+        let mut websocket = common::handle_websocket(&server);
 
-        handle_config_request_enterprise_example(&server);
+        common::handle_config_request_enterprise_example(&server);
 
         // Wait for the client to recieve (and test) the first config
         update_config_rx.recv().unwrap();
@@ -77,24 +39,6 @@ fn server_thread() -> ServerHandle {
         _terminator: terminator,
         config_updated: config_updated_tx,
         port,
-    }
-}
-
-#[derive(Debug)]
-struct MockTokenProvider {}
-
-impl TokenProvider for MockTokenProvider {
-    fn get_access_token(&self) -> appconfiguration::NetworkResult<String> {
-        Ok("mock_token".into())
-    }
-}
-
-fn wait_until_online(client: &AppConfigurationClientHttp<LiveConfigurationImpl>) {
-    loop {
-        if client.is_online().unwrap() {
-            break;
-        };
-        sleep(Duration::from_millis(10));
     }
 }
 
@@ -113,10 +57,10 @@ fn main() {
         "collection_id".to_string(),
     );
     let client =
-        AppConfigurationClientHttp::new(address, Box::new(MockTokenProvider {}), config_id)
+        AppConfigurationClientHttp::new(address, Box::new(common::MockTokenProvider {}), config_id)
             .unwrap();
 
-    wait_until_online(&client);
+    common::wait_until_online(&client);
 
     // Tell the server that now it can progress
     server.config_updated.send(()).unwrap();
